+++ rust/main/chains/hyperlane-cosmos/src/mailbox/contract.rs
@@ -41,7 +41,7 @@ impl CosmosMailbox {
         let provider = CosmosProvider::new(
             locator.domain.clone(),
             conf.clone(),
-            Some(locator.clone()),
+            locator.clone(),
             signer,
         )?;