+++ typescript/infra/config/environments/mainnet3/igp.ts
@@ -1,19 +1,10 @@
-import { BigNumber, ethers } from 'ethers';
-
-import {
-  ChainMap,
-  ChainName,
-  HookType,
-  IgpConfig,
-  TOKEN_EXCHANGE_RATE_DECIMALS,
-  defaultMultisigConfigs,
-  multisigIsmVerificationCost,
-} from '@hyperlane-xyz/sdk';
+import { ChainMap, ChainName, HookType, IgpConfig } from '@hyperlane-xyz/sdk';
 import { exclude, objMap } from '@hyperlane-xyz/utils';
 
 import {
   AllStorageGasOracleConfigs,
   getAllStorageGasOracleConfigs,
+  getOverhead,
   getTokenExchangeRateFromValues,
 } from '../../../src/config/gas-oracle.js';
 
@@ -25,37 +16,22 @@ import rawTokenPrices from './tokenPrices.json';
 
 const tokenPrices: ChainMap<string> = rawTokenPrices;
 
-const FOREIGN_DEFAULT_OVERHEAD = 600_000; // cosmwasm warp route somewhat arbitrarily chosen
-
-const remoteOverhead = (remote: ChainName) =>
-  ethereumChainNames.includes(remote as any)
-    ? multisigIsmVerificationCost(
-        defaultMultisigConfigs[remote].threshold,
-        defaultMultisigConfigs[remote].validators.length,
-      )
-    : FOREIGN_DEFAULT_OVERHEAD; // non-ethereum overhead
-
-// Gets the exchange rate of the remote quoted in local tokens
-function getTokenExchangeRate(local: ChainName, remote: ChainName): BigNumber {
-  const localValue = ethers.utils.parseUnits(
-    tokenPrices[local],
-    TOKEN_EXCHANGE_RATE_DECIMALS,
-  );
-  const remoteValue = ethers.utils.parseUnits(
-    tokenPrices[remote],
-    TOKEN_EXCHANGE_RATE_DECIMALS,
-  );
-
-  return getTokenExchangeRateFromValues(local, localValue, remote, remoteValue);
-}
+const getOverheadWithOverrides = (local: ChainName, remote: ChainName) => {
+  let overhead = getOverhead(local, remote, ethereumChainNames);
+  if (remote === 'moonbeam') {
+    overhead *= 4;
+  }
+  return overhead;
+};
 
 const storageGasOracleConfig: AllStorageGasOracleConfigs =
   getAllStorageGasOracleConfigs(
     supportedChainNames,
     gasPrices,
-    getTokenExchangeRate,
+    (local, remote) =>
+      getTokenExchangeRateFromValues(local, remote, tokenPrices),
     (local) => parseFloat(tokenPrices[local]),
-    (local) => remoteOverhead(local),
+    getOverheadWithOverrides,
   );
 
 export const igp: ChainMap<IgpConfig> = objMap(
@@ -73,7 +49,7 @@ export const igp: ChainMap<IgpConfig> = objMap(
     overhead: Object.fromEntries(
       exclude(local, supportedChainNames).map((remote) => [
         remote,
-        remoteOverhead(remote),
+        getOverheadWithOverrides(local, remote),
       ]),
     ),
     oracleConfig: storageGasOracleConfig[local],