+++ typescript/sdk/src/consts/multisigIsm.ts
@@ -80,6 +80,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     validators: ['0x82e3b437a2944e3ff00258c93e72cd1ba5e0e921'],
   },
 
+  berabartio: {
+    threshold: 1,
+    validators: ['0x541dd3cb282cf869d72883557badae245b63e1fd'],
+  },
+
   bitlayer: {
     threshold: 2,
     validators: [
@@ -126,6 +131,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
+  camptestnet: {
+    threshold: 1,
+    validators: ['0x238f40f055a7ff697ea6dbff3ae943c9eae7a38e'],
+  },
+
   celo: {
     threshold: 3,
     validators: [
@@ -152,6 +162,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x0874967a145d70b799ebe9ed861ab7c93faef95a',
       '0xd767ea1206b8295d7e1267ddd00e56d34f278db6',
     ],
+  },
+
+  citreatestnet: {
+    threshold: 1,
+    validators: ['0x60d7380a41eb95c49be18f141efd2fde5e3dba20'],
   },
 
   connextsepolia: {
@@ -255,6 +270,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
+  formtestnet: {
+    threshold: 1,
+    validators: ['0x72ad7fddf16d17ff902d788441151982fa31a7bc'],
+  },
+
   fraxtal: {
     threshold: 2,
     validators: [
@@ -295,6 +315,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   holesky: {
     threshold: 1,
     validators: ['0x7ab28ad88bb45867137ea823af88e2cb02359c03'], // TODO
+  },
+
+  hyperliquidevmtestnet: {
+    threshold: 1,
+    validators: ['0xea673a92a23ca319b9d85cc16b248645cd5158da'],
   },
 
   inevm: {
@@ -609,6 +634,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     validators: ['0xd4ce8fa138d4e083fc0e480cca0dbfa4f5f30bd5'],
   },
 
+  soneiumtestnet: {
+    threshold: 1,
+    validators: ['0x2e2101020ccdbe76aeda1c27823b0150f43d0c63'],
+  },
+
   stride: {
     threshold: 6,
     validators: [
@@ -623,6 +653,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x5937b7cE1029C3Ec4bD8e1AaCc0C0f9422654D7d', // stakecito
       '0x3a446ed2923c08445af06e53f0acb558c0e0413c', // staked
     ],
+  },
+
+  suavetoliman: {
+    threshold: 1,
+    validators: ['0xf58f6e30aabba34e8dd7f79b3168507192e2cc9b'],
   },
 
   superpositiontestnet: {