+++ typescript/github-proxy/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @hyperlane-xyz/github-proxy
 
+## 5.3.0
+
+## 5.2.1
+
 ## 5.2.0
 
 ### Minor Changes