+++ typescript/helloworld/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @hyperlane-xyz/helloworld
 
+## 5.3.0
+
+### Minor Changes
+
+- 35d4503b9: Update to registry v4.3.6
+
+### Patch Changes
+
+- Updated dependencies [eb47aaee8]
+- Updated dependencies [50319d8ba]
+- Updated dependencies [8de531fa4]
+- Updated dependencies [fd536a79a]
+  - @hyperlane-xyz/sdk@5.3.0
+  - @hyperlane-xyz/core@5.3.0
+
+## 5.2.1
+
+### Patch Changes
+
+- Updated dependencies [eb5afcf3e]
+  - @hyperlane-xyz/core@5.2.1
+  - @hyperlane-xyz/sdk@5.2.1
+
 ## 5.2.0
 
 ### Minor Changes